GHCP CS4.5: fix styling, add hashtag feature during the refactor to quasar, the neo-brutalism theme broke. please use this opportunity to rework the whole ui to a whatsapp/signal/messenger like UI but with neo-brutalism, you are allowed to break the current functionality to revamp it to better ui/ux. refactor the modules if needed, and use e.g. frameworks e.g. shadcn and unocss to your advantage, check your package.json and codebase to determine
there are still styling issues e.g. alignment and lack of support for advanced features e.g. hashtag tagging too, please fix and take notes from ui/ux design feedback on neo-brutalism and design principles

please ensure that hashtags can be easily added and deleted too by clicking a button, and it'll remove the text too, and ensure that the payload when saving will also capture the hashtags separately as like a set. please also implement note editing functionality, and incorporate the same hashtag functionality too. hashtags should be searchable too and approrpriate ui/ux considerations be done too, and consider this to implement a sidebar integrating a hashtag hierarchy too for the next step once the above and related tasks are done.

the hashtag functionality doesn't work with unicode characters. also, it is confusing when we remove hashtags, it also changes the note content too, we should separate the hashtags and content for good ui/ux, especially on editing (e.g. we could simply just have hashtag manipulating the text as a configurable helper that only applies on the creation and not editing), and we should also keep in mind the edit to have edit history too. edits should leverage the crdt nature of tiptap if posslble, if you haven't noticed that you have this installed already (and gracefully degrade to current behavior if not available). the search should have better ui/ux to support hashtags too

the code is too convoluted that way. instead, we should simply prompt the user to delete all notes or do a one off migration, or at least modularize the code to make it explicitly and allow for easy deprecation in the future. go through the entire codebase to refactor in similar fashion too. in addition, the toggle to auto-extract tags still doesn't work for editing notes, editing notes should by default not edit the content too, please revise and make that a separate togglable function if possible too